Update MIDI Interface to RT Light Version
## Goal
Remove Sysex handling and per message type callbacks to save code
size on flash-constrained AVR project

## Implementation

### Changes to MidiInterface class definition

#### Removed _Settings and _Transport from the template.
Only DefaultSettings and Arduino platform will be supported by the
Light version.

#### Removed SysEx functions
Removed sendSysEx(), getSysExArray(), and getSysExArrayLength() as
sysex is not supported.  Also removed non-class functions
encodeSysEx() and decodeSysEx() for the same reason

#### Removed Individual Message Type Callbacks
Processing individual message callbacks for every received message
adds a lot of unnecessary code. Moving this functionality to the
consumer of the library allows that firmware to determine its
complexity and saves a lot of code space for a project only using
Program Change and Control Change messages

#### Simplify Running Status Support
Removed running status caches for TX and RX and replaced them with
one running status buffer byte since running status will not be used
on transmission to save code space.  The light version is designed
for non-instrument devices so running status isn't necessary at that
message density.

#### Removed RPN Support
RPN memory buffers and handling are moved

#### Removed Timestamps
MIDI Clock and active sensing will not be supported by this library
so timestamps are not necessary.  Any timing should be performed in
the consumers MessageCallback

#### Added parseLight() and Support Functions
Incoming message parsing will now use a new implementation called
parseLight(). Added the following support functions to help with
parsing and make the method more legible:
    - preparePendingMessage
    - handleRealTimeByte
    - handleDataByte
    - isStatusByte
    - isRealTimeByte

#### Changed Header File Names and Namespace Reference to MIDI Light
- MIDI_NAMESPACE > MIDI_LIGHT_NAMESPACE
- MIDI.h > MIDI_RTLight.h
- MIDI.hpp > MIDI_RTLight.hpp

### Changes to MidiInterface class implementation

##### Commented Out all References to Removed Functionality
See 'Changes to MidiInterface class definition'.

##### Updated Template Functions to remove_Settings and _Transport

##### Simplified read() function
- Removed unsupported features (active sensing, null note velocity)
- Changed parsing method to parseLight and moved transport reading
  to the read() function and parseLight parses only a single byte
- Moved thruFiler up so that it can have slightly lower latency

##### Added parseLight() function
parseLight parses the incoming byte.  Valid bytes for channel messages
with or without running status are handled.  Realtime messages and
other system common message are ignored.  This functions returns true
anytime a valid MIDI Message has been completely received

##### Added handleStatusByte() function
If the status byte is a channel message type, it stores it in memory
and determined the expected message length.  If it is not a channel
message, the byte is discarded and status is marked invalid.

##### Added handleDataByte() function
This functions receives data bytes.  It managed running status and
stores data bytes as expected.  Once the last expected data byte is
received, it returns true.  If an unexpected data byte is received
when there is not valid running status, it returns false

##### Added preparePendingMessage() function
This functions populates the pending message structure with the valid
data stored in the receive parsing data buffers

##### Simplified lanchCallback() function
Since specific message type callbacks are removed, this function now
simply checks that the mMessageCallback member as been set and then
calls it if it has

##### Simplified thruFilter() function
Thru filter function only supports control change and program change
messages for code size optimization